Day 9/startup-marketability-evaluator/requirements.txt
langchain==0.1.0
google-generativeai==0.4.1
chromadb==0.4.22
streamlit==1.32.0
pandas==2.1.4
python-dotenv==1.0.0
pytrends==4.9.0
serpapi==0.1.0
langchain-google-genai==0.0.11